allow empty plots to be manipulated
This commit fixes issue #733 where uninitialized axes (plots with no data in them) cannot be dragged to pan or zoom.

We don't want to initialize them immediately when a new Plot() is created, because uninitialized axes are an indication that AxisAuto() should be called automatically before the first Render().

Instead, this change checks if axes are uninitialized when the mouse interacts with them, and if so it sets their limits (to what they already are, the defaults) marking them as set so they can be manipulated by panning and zooming.