Fix - `MouseZoomRect` affects all axes
This PR fixes #1537 to preserve the aspect ratio between axes during a Zoom.
In fact, the current behavior seems in contradiction with the fact that by default axes limits are free and should (to my understanding) be affected in the same manner by the drag zoom, regardless of the axis.