GDI: Rename dangerous MeasureString()
One ScottPlot.Drawing.MeasureString() overload creates a temporary Bitmap and Graphics just to measure a string. To prevent this overload from being accidentally called instead of a more performant one, it is being renamed to MeasureStringUsingTemporaryGraphics() which is more expressive.

Callers that were using it by accident were updated to point to a MeasureString() that uses an existing Graphics.

XML docs were added to MeasureString() overloads to clarify their behavior.