SP4 Tests: Histogram count and bin lengths
Assert that the new Histogram class always returns a number of bin edges equal to the number of counts.

Note that this is *NOT* the behavior of other numerical libraries like numpy, but doing it this way makes it easier to create X/Y and bar plots since the number of Xs will match the number of Ys.