WPF: fix XAML component naming
Always use `x:Name` for naming components in XAML code.

This is to most "common" way to name components in XAML/WPF as the XAML compiler will ensure this name will be unique and it cannot be changed later in code behind via `SetValue()`.
This way, it avoids any naming conflicts.